+++ src/bluer/characteristic.rs
@@ -100,6 +100,25 @@ impl CharacteristicImpl {
             .await;
     }
 
+    /// Get the maximum amount of data that can be written in a single packet for this characteristic.
+    pub fn max_write_len(&self) -> Result<usize> {
+        // Call an async function from a synchronous context
+        match tokio::runtime::Handle::try_current() {
+            Ok(handle) => tokio::task::block_in_place(move || handle.block_on(self.max_write_len_async())),
+            Err(_) => tokio::runtime::Builder::new_current_thread()
+                .build()
+                .unwrap()
+                .block_on(self.max_write_len_async()),
+        }
+    }
+
+    /// Get the maximum amount of data that can be written in a single packet for this characteristic.
+    pub async fn max_write_len_async(&self) -> Result<usize> {
+        let mtu = self.inner.mtu().await?;
+        // GATT characteristic writes have 3 bytes of overhead (opcode + handle id)
+        Ok(mtu - 3)
+    }
+
     /// Enables notification of value changes for this GATT characteristic.
     ///
     /// Returns a stream of values for the characteristic sent from the device.
@@ -109,7 +128,7 @@ impl CharacteristicImpl {
 
     /// Is the device currently sending notifications for this characteristic?
     pub async fn is_notifying(&self) -> Result<bool> {
-        self.inner.notifying().await.map_err(Into::into)
+        Ok(self.inner.notifying().await?.unwrap_or(false))
     }
 
     /// Discover the descriptors associated with this characteristic.
@@ -119,8 +138,7 @@ impl CharacteristicImpl {
 
     /// Get previously discovered descriptors.
     ///
-    /// If no descriptors have been discovered yet, this method may either perform descriptor discovery or
-    /// return an error.
+    /// If no descriptors have been discovered yet, this method will perform descriptor discovery.
     pub async fn descriptors(&self) -> Result<Vec<Descriptor>> {
         self.inner
             .descriptors()