+++ src/bluer/service.rs
@@ -71,8 +71,7 @@ impl ServiceImpl {
 
     /// Get previously discovered characteristics.
     ///
-    /// If no characteristics have been discovered yet, this method may either perform characteristic discovery or
-    /// return an error.
+    /// If no characteristics have been discovered yet, this method will perform characteristic discovery.
     pub async fn characteristics(&self) -> Result<Vec<Characteristic>> {
         self.inner
             .characteristics()
@@ -93,8 +92,7 @@ impl ServiceImpl {
 
     /// Get previously discovered included services.
     ///
-    /// If no included services have been discovered yet, this method may either perform included service discovery
-    /// or return an error.
+    /// If no included services have been discovered yet, this method will perform included service discovery.
     pub async fn included_services(&self) -> Result<Vec<Service>> {
         let session = super::adapter::session().await?;
         let adapter = session.adapter(self.inner.adapter_name())?;